requirements.txt
Flask==3.0.0
openai==1.2.3
requests==2.31.0
apify==1.2.0
beautifulsoup4==4.12.2
alpha-vantage==2.3.1
pandas
asyncio
flask-cors
python-dotenv
